Remove unwanted affiliate toast
There was no need for that annoying advertisement toast. Adding this advertisement only annoyed more users than it did make them click on that link. Yes, we know, you may rely on this affiliation in one way or another. Putting it in your README is fine - users will click on the course at their own accord. But subjecting all your regular, everyday users of this extension (myself included) to this kindof unwanted advertising - big no-no. So I suggest you accept this PR to minimise the damage done to your extension.